Move the incremental testing down to the SourceGeneratorTest type.
