+++ src/Microsoft.CodeAnalysis.Testing/Microsoft.CodeAnalysis.Analyzer.Testing/AnalyzerTest`1.cs
@@ -18,7 +18,6 @@ using DiffPlex;
 using DiffPlex.Chunkers;
 using DiffPlex.DiffBuilder;
 using DiffPlex.DiffBuilder.Model;
-using Humanizer;
 using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis.Host.Mef;
 using Microsoft.CodeAnalysis.Options;
@@ -81,7 +80,6 @@ namespace Microsoft.CodeAnalysis.Testing
         protected AnalyzerTest()
         {
             TestState = new SolutionState(DefaultTestProjectName, Language, DefaultFilePathPrefix, DefaultFileExt);
-            IncrementalChangeTestState = new SolutionState(DefaultTestProjectName, Language, DefaultFilePathPrefix, DefaultFileExt);
         }
 
         /// <summary>
@@ -125,8 +123,6 @@ namespace Microsoft.CodeAnalysis.Testing
         public MarkupOptions MarkupOptions { get; set; }
 
         public SolutionState TestState { get; }
-
-        public SolutionState IncrementalChangeTestState { get; }
 
         /// <summary>
         /// Gets the collection of inputs to provide to the XML documentation resolver.
@@ -174,11 +170,6 @@ namespace Microsoft.CodeAnalysis.Testing
         public List<Func<Solution, ProjectId, Solution>> SolutionTransforms { get; } = new List<Func<Solution, ProjectId, Solution>>();
 
         /// <summary>
-        /// Gets the expected state for incremental generator states per generator type.
-        /// </summary>
-        public Dictionary<Type, IncrementalGeneratorExpectedState> IncrementalGeneratorStepStates { get; } = new Dictionary<Type, IncrementalGeneratorExpectedState>();
-
-        /// <summary>
         /// Gets or sets the timeout to use when matching expected and actual diagnostics. The default value is 2
         /// seconds.
         /// </summary>
@@ -223,14 +214,9 @@ namespace Microsoft.CodeAnalysis.Testing
             var defaultDiagnostic = GetDefaultDiagnostic(analyzers);
             var supportedDiagnostics = analyzers.SelectMany(analyzer => analyzer.SupportedDiagnostics).ToImmutableArray();
             var fixableDiagnostics = ImmutableArray<string>.Empty;
+            var testState = TestState.WithInheritedValuesApplied(null, fixableDiagnostics).WithProcessedMarkup(MarkupOptions, defaultDiagnostic, supportedDiagnostics, fixableDiagnostics, DefaultFilePath);
 
-            var rawTestState = TestState.WithInheritedValuesApplied(null, fixableDiagnostics);
-            var rawIncrementalTestState = IncrementalChangeTestState.WithInheritedValuesApplied(rawTestState, fixableDiagnostics);
-
-            var testState = rawTestState.WithProcessedMarkup(MarkupOptions, defaultDiagnostic, supportedDiagnostics, fixableDiagnostics, DefaultFilePath);
-            var incrementalTestState = rawIncrementalTestState.WithProcessedMarkup(MarkupOptions, defaultDiagnostic, supportedDiagnostics, fixableDiagnostics, DefaultFilePath);
-
-            var diagnostics = await VerifySourceGeneratorAsync(testState, incrementalTestState, Verify, cancellationToken).ConfigureAwait(false);
+            var diagnostics = await VerifySourceGeneratorAsync(testState, Verify, cancellationToken).ConfigureAwait(false);
             await VerifyDiagnosticsAsync(new EvaluatedProjectState(testState, ReferenceAssemblies), testState.AdditionalProjects.Values.Select(additionalProject => new EvaluatedProjectState(additionalProject, ReferenceAssemblies)).ToImmutableArray(), testState.ExpectedDiagnostics.ToArray(), Verify, cancellationToken).ConfigureAwait(false);
         }
 
@@ -285,11 +271,10 @@ namespace Microsoft.CodeAnalysis.Testing
         /// Called to test a C# source generator when applied on the input source as a string.
         /// </summary>
         /// <param name="testState">The effective input test state.</param>
-        /// <param name="incrementalTestState">The effective test state for the incremental generator test.</param>
         /// <param name="verifier">The verifier to use for test assertions.</param>
         /// <param name="cancellationToken">The <see cref="CancellationToken"/> that the task will observe.</param>
         /// <returns>A <see cref="Task"/> representing the asynchronous operation.</returns>
-        protected async Task<ImmutableArray<Diagnostic>> VerifySourceGeneratorAsync(SolutionState testState, SolutionState incrementalTestState, IVerifier verifier, CancellationToken cancellationToken)
+        protected async Task<ImmutableArray<Diagnostic>> VerifySourceGeneratorAsync(SolutionState testState, IVerifier verifier, CancellationToken cancellationToken)
         {
             var sourceGenerators = GetSourceGenerators().ToImmutableArray();
             if (sourceGenerators.IsEmpty)
@@ -297,14 +282,14 @@ namespace Microsoft.CodeAnalysis.Testing
                 return ImmutableArray<Diagnostic>.Empty;
             }
 
-            return await VerifySourceGeneratorAsync(Language, sourceGenerators, testState, incrementalTestState, verifier.PushContext("Source generator application"), cancellationToken);
+            var driver = await VerifySourceGeneratorAsync(Language, sourceGenerators, testState, verifier.PushContext("Source generator application"), cancellationToken);
+            return driver.GetRunResult().Diagnostics;
         }
 
-        private protected async Task<ImmutableArray<Diagnostic>> VerifySourceGeneratorAsync(
+        private protected async Task<LightupGeneratorDriver> VerifySourceGeneratorAsync(
             string language,
             ImmutableArray<Type> sourceGenerators,
             SolutionState testState,
-            SolutionState incrementalTestState,
             IVerifier verifier,
             CancellationToken cancellationToken)
         {
@@ -356,34 +341,7 @@ namespace Microsoft.CodeAnalysis.Testing
                     MatchDiagnosticsTimeout);
             }
 
-            // If any solution transforms for incremental generator testing have been defined,
-            // then test incremental generators.
-            if (IncrementalGeneratorStepStates.Count != 0)
-            {
-                var secondRunProject = await CreateProjectAsync(new EvaluatedProjectState(incrementalTestState, ReferenceAssemblies), incrementalTestState.AdditionalProjects.Values.Select(additionalProject => new EvaluatedProjectState(additionalProject, ReferenceAssemblies)).ToImmutableArray(), cancellationToken);
-
-                updatedDriver = updatedDriver
-                        .ReplaceAdditionalTexts(secondRunProject.AnalyzerOptions.AdditionalFiles)
-                        .WithUpdatedParseOptions(secondRunProject.ParseOptions)
-                        .WithUpdatedAnalyzerConfigOptions(secondRunProject.AnalyzerOptions.AnalyzerConfigOptionsProvider());
-
-                var (_, secondRunDriver) = await ApplySourceGeneratorsAsync(secondRunProject, updatedDriver, verifier, cancellationToken).ConfigureAwait(false);
-
-                var generatorRunResults = secondRunDriver.GetRunResult();
-
-                foreach (var generatorResult in generatorRunResults.Results)
-                {
-                    if (generatorResult.GeneratorType is not { } generatorType
-                        || generatorResult.TrackedSteps is not { } trackedSteps)
-                    {
-                        continue;
-                    }
-
-                    VerifyGeneratorSteps(verifier.PushContext("Verifying source generator incremental step state"), generatorType, trackedSteps);
-                }
-            }
-
-            return diagnostics;
+            return generatorDriver;
 
             static void VerifyDocuments(
                 IVerifier verifier,
@@ -489,51 +447,6 @@ namespace Microsoft.CodeAnalysis.Testing
                     var (fileName, folders) = getNameAndFolders(defaultFilePathPrefix, expected.filename);
                     verifier.Equal(fileName, actual.document.Name, $"file name was expected to be '{fileName}' but was '{actual.document.Name}'");
                     verifier.SequenceEqual(folders, actual.document.Folders, message: $"folders was expected to be '{string.Join("/", folders)}' but was '{string.Join("/", actual.document.Folders)}'");
-                }
-            }
-
-            void VerifyGeneratorSteps(IVerifier verifier, Type generatorType, ImmutableDictionary<string, ImmutableArray<LightupIncrementalGeneratorRunStep>> trackedSteps)
-            {
-                if (IncrementalGeneratorStepStates.TryGetValue(generatorType, out var expectedState))
-                {
-                    var expectedStepNames = new HashSet<string>(expectedState.ExpectedStepStates.Keys);
-
-                    expectedStepNames.ExceptWith(trackedSteps.Keys);
-
-                    if (expectedStepNames.Count != 0)
-                    {
-                        verifier.Fail($"Expected to see steps with the following names, but they were not executed by the '{generatorType.FullName}' generator: {string.Join(", ", expectedStepNames)}");
-                    }
-
-                    foreach (var trackedStep in trackedSteps)
-                    {
-                        if (expectedState.ExpectedStepStates.TryGetValue(trackedStep.Key, out var expectedStepStates))
-                        {
-                            var trackedStepExecutions = trackedStep.Value;
-                            verifier.Equal(expectedStepStates.Count, trackedStepExecutions.Length, $"Expected {expectedStepStates.Count} executions of step '{trackedStep.Key}' but there were only {trackedStepExecutions.Length} executions");
-                            for (var i = 0; i < expectedStepStates.Count; i++)
-                            {
-                                var stepNumber = i + 1;
-                                verifier.Equal(expectedStepStates[i].InputRunReasons.Count, trackedStepExecutions[i].Inputs.Length, $"Expected {expectedStepStates[i].InputRunReasons.Count} inputs for the '{trackedStep.Key}' step's {stepNumber.Ordinalize()} execution but there was {"input".ToQuantity(trackedStepExecutions[i].Inputs.Length)}");
-                                for (var j = 0; j < expectedStepStates[i].InputRunReasons.Count; j++)
-                                {
-                                    var inputNumber = j + 1;
-                                    var expectedInputState = expectedStepStates[i].InputRunReasons[j];
-                                    var actualInputState = trackedStepExecutions[i].Inputs[j].Input.Outputs[trackedStepExecutions[i].Inputs[j].OutputIndex].Reason;
-                                    verifier.Equal(expectedInputState, actualInputState, $"Expected the {inputNumber.Ordinalize()} input state for the {stepNumber.Ordinalize()} '{trackedStep.Key}' step to be '{expectedInputState}' but it was '{actualInputState}'");
-                                }
-
-                                verifier.Equal(expectedStepStates[i].OutputRunReasons.Count, trackedStepExecutions[i].Outputs.Length, $"Expected {expectedStepStates[i].OutputRunReasons.Count} outputs for the '{trackedStep.Key}' step's {stepNumber.Ordinalize()} execution but there was {"outputs".ToQuantity(trackedStepExecutions[i].Outputs.Length)}");
-                                for (var j = 0; j < expectedStepStates[i].OutputRunReasons.Count; j++)
-                                {
-                                    var outputNumber = j + 1;
-                                    var expectedOutputState = expectedStepStates[i].OutputRunReasons[j];
-                                    var actualOutputState = trackedStepExecutions[i].Outputs[j].Reason;
-                                    verifier.Equal(expectedOutputState, actualOutputState, $"Expected the {outputNumber.Ordinalize()} output state for the {stepNumber.Ordinalize()} '{trackedStep.Key}' step to be '{expectedOutputState}' but it was '{actualOutputState}'");
-                                }
-                            }
-                        }
-                    }
                 }
             }
         }
@@ -1404,19 +1317,36 @@ namespace Microsoft.CodeAnalysis.Testing
 
             // If we cannot specify generator options or if we are not testing incrementality
             // don't try to track incremental steps.
-            if (generatorDriverOptionsType is null || IncrementalGeneratorStepStates.Count == 0)
+            if (generatorDriverOptionsType is null)
             {
                 driver = createMethod.Invoke(null, new[] { convertedSourceGenerators, additionalFiles, project.ParseOptions, analyzerConfigOptionsProvider });
             }
             else
             {
-                var constructor = generatorDriverOptionsType.GetConstructor(
+                var constructorWithIncrementalTracking = generatorDriverOptionsType.GetConstructor(
                     new[]
                     {
                         generatorDriverOptionsType.GetTypeInfo().Assembly.GetType("Microsoft.CodeAnalysis.IncrementalGeneratorOutputKind")!,
                         typeof(bool),
                     });
-                var options = constructor!.Invoke(new object[] { 0, true });
+
+                object options;
+
+                if (constructorWithIncrementalTracking is not null)
+                {
+                    options = constructorWithIncrementalTracking.Invoke(new object[] { 0, true });
+                }
+                else
+                {
+                    var constructorWithOnlyOutputKinds = generatorDriverOptionsType.GetConstructor(
+                        new[]
+                        {
+                            generatorDriverOptionsType.GetTypeInfo().Assembly.GetType("Microsoft.CodeAnalysis.IncrementalGeneratorOutputKind")!,
+                        });
+                    verifier.True(constructorWithOnlyOutputKinds is not null);
+                    options = constructorWithOnlyOutputKinds!.Invoke(new object[] { 0 });
+                }
+
                 driver = createMethodWithOptions!.Invoke(null, new[] { convertedSourceGenerators, additionalFiles, project.ParseOptions, analyzerConfigOptionsProvider, options });
             }
